requirements.txt
## Models
scikit-learn
torch==2.4.0
torchvision==0.19.0
xgboost==2.1.1
tensorboardx==2.6.2.2

## Others
cython==3.0.11
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
pyyaml==6.0.1
packaging==24.1
pycocotools==2.0.8
h5py==3.11.0
requests==2.32.3
tqdm==4.66.5
seaborn==0.13.2

ensure==1.0.4 
python-box==7.2.0
python-dotenv==1.0.1


## MLOps tools
dvc[all]==3.56.0
dagshub==0.3.40
mlflow==2.17.1
zenml==0.68.1

## docs
mkdocs-material==9.5.42
kaggle==1.6.17 

## logs and alerting
opentelemetry-distro==0.48b0
opentelemetry-exporter-otlp==1.27.0
python-logging-loki


## dev 
### For Dev Environment

pytest==8.3.3
tox==4.21.0
mypy==1.11.2
pytest-cov==5.0.0
flake8>=5.0.4